feat(runtime-utils): split parsed from browser
