+++ .changeset/afraid-peaches-chew.md
@@ -1,0 +1,8 @@
+---
+'@modern-js/plugin-router-v5': patch
+'@modern-js/runtime': patch
+'@modern-js/runtime-utils': patch
+---
+
+feat(runtime-utils): split parsed from browser
+feat(runtime-utils): 从 browser 中拆分 parsed 方法
