+++ packages/runtime/plugin-runtime/src/router/runtime/plugin.tsx
@@ -9,7 +9,7 @@ import {
   RouteObject,
 } from '@modern-js/runtime-utils/router';
 import hoistNonReactStatics from 'hoist-non-react-statics';
-import { parsedJSONFromElement } from '@modern-js/runtime-utils/browser';
+import { parsedJSONFromElement } from '@modern-js/runtime-utils/parsed';
 import type { RouterSubscriber } from '@modern-js/runtime-utils/remix-router';
 import { Plugin, RuntimeReactContext } from '../../core';
 import { modifyRoutes as modifyRoutesHook } from './hooks';
