+++ packages/toolkit/runtime-utils/src/browser/index.ts
@@ -1,2 +1,1 @@
-export * from './parsed';
 export * from './nestedRoutes';
