+++ packages/toolkit/runtime-utils/tests/brower/parse.test.ts
@@ -1,4 +1,4 @@
-import { parsedJSONFromElement } from '../../src/browser';
+import { parsedJSONFromElement } from '../../src/parsed';
 
 describe('parsed utils', () => {
   it('should return null if no alias config', () => {
